Separated cli and server logic
