+++ bin/src/filesystem/models/folder.model.ts
@@ -10,7 +10,9 @@ export class Folder {
   }
 
   async create(dirpath: PathLike): Promise<string | undefined> {
-    return await fs.mkdir(path.join(String(dirpath), this.name), { recursive: true });
+    return await fs.mkdir(path.join(String(dirpath), this.name), {
+      recursive: true,
+    });
   }
 
   static async list(dirpath: PathLike): Promise<string[]> {
