+++ test/e2e/app.e2e-spec.ts
@@ -1,11 +1,11 @@
 import dotenv from 'dotenv';
 import supertest from 'supertest';
-import { app, server } from '../../bin';
 import { UserDto } from '../../bin/src/auth/dto';
 import { db, tableName } from '../../bin/db';
 import * as fs from 'fs';
 import path from 'path';
 import { Folder, storageRoot } from '../../bin/src/filesystem';
+import { app, server } from '../../bin/src/server';
 
 dotenv.config();
